chore: config restructuring
